app/requirements.txt
fastapi
uvicorn
requests
textblob
python-dotenv
pandas